the last of the compiler warning fixes for the luatexdir subdirectories (only dvi/font/image/lang/lua/ocp/pdf/tex/utils), and indent
git-svn-id: https://serveur-svn.lri.fr/svn/modhel/luatex/trunk@3404 0b2b3880-5936-4365-a048-eb17d2e5a6bf

+++ source/texk/web2c/luatexdir/font/sfnt.h
@@ -181,8 +181,8 @@ typedef struct {
 
 #    define ASSERT(a) assert(a)
 #    define RELEASE(a) free(a)
-#    define NEW(a,b) xmalloc((a)*sizeof(b))
-#    define RENEW(a,b,c) xrealloc(a, ((b)*sizeof(c)))
+#    define NEW(a,b) xmalloc((unsigned)((unsigned)(a)*sizeof(b)))
+#    define RENEW(a,b,c) xrealloc(a, (unsigned)((unsigned)(b)*sizeof(c)))
 #    define TT_ERROR pdftex_fail
 #    define WARN printf
